hugging-face-deployment/requirements.txt
torch
torchvision
pillow
gradio
matplotlib
pandas
